chore(custom): cli
ability to include comma seperated list
